allow pick to selectively ignore parent fields
The `pick()` operator is limited by the fact that `class-transformer` and `class-validator`
both obscure their decorator metadata, meaning that we cannot ask a class how it is decorated,
we can only modify its decoration via **additional** decorators.

While `dto-decorators` promises an alternative (because it supports saving decorator metadata
in our own storage), this solution-path is a bit longer than simply modifying the decorator
logic to ignore certain fields.

That is:

 - The return type to `pick()` is computed based on the provided `fields`.
 - The default behavior is that we mark every non-included field as either `@Exclude` or `@IsOptional`
 - The new behavior allows us to ignore certain fields that are not part of the returned type
   so we can extend them subsequently.